+++ src/to-static-html/url-as-html.js
@@ -15,8 +15,7 @@ export default (url, { metadata, routes, store, baseUrl }) => (
       `<meta charset="utf-8" />`,
       `<meta http-equiv="X-UA-Compatible" content="IE=edge" />`,
       `<meta name="viewport" content="width=device-width, initial-scale=1" />`,
-      `<link rel="stylesheet" href="${
-        baseUrl.path }/statinamic-client.css" />`,
+      `<link rel="stylesheet" href="${ baseUrl.path }statinamic-client.css" />`,
     ].join("")
 
     try {
@@ -100,7 +99,7 @@ export default (url, { metadata, routes, store, baseUrl }) => (
                 body,
                 script,
                 children: (
-                  <script src={ baseUrl.path + "/statinamic-client.js" }>
+                  <script src={ `${ baseUrl.path }statinamic-client.js` }>
                   </script>
                 ),
               })